27-cellular-connected-electronic-kiosk/requirements.txt
note_python==1.4.6
python_periphery==2.3.0
